Add ruff with configuration for math code and auto-fix issues
- Add ruff >=0.8.6 to pixi dependencies
- Add ruff tasks: ruff-check, ruff-fix, ruff-format
- Configure ruff to ignore uppercase variable names (N803, N806, N812)
  - Allows X, W, Wx, Wh, etc. for mathematical notation
  - Common in ML/scientific code
- Auto-fixed 801 linting issues:
  - Removed 769 blank lines with whitespace
  - Fixed 44 f-strings missing placeholders
  - Removed 26 unused imports
  - Sorted 24 import statements
- Formatted 35 files with ruff format

Remaining: 23 minor issues (bare excepts, unused vars, import order)

+++ src/python/max_distilbert/embeddings.py
@@ -17,7 +17,7 @@ class DistilBertEmbeddings(Module):
         device: DeviceRef,
     ):
         """Initialize DistilBERT embeddings.
-        
+
         Args:
             weights: Embedding weights from the model
             config: HuggingFace model configuration
@@ -25,42 +25,36 @@ class DistilBertEmbeddings(Module):
             device: Device to run on
         """
         super().__init__()
-        
+
         # Word embeddings
-        self.word_embeddings = weights.word_embeddings.weight.allocate(
-            DType.float32
-        ).cast(dtype)
-        
+        self.word_embeddings = weights.word_embeddings.weight.allocate(DType.float32).cast(dtype)
+
         # Position embeddings
-        self.position_embeddings = weights.position_embeddings.weight.allocate(
-            DType.float32
-        ).cast(dtype)
-        
+        self.position_embeddings = weights.position_embeddings.weight.allocate(DType.float32).cast(
+            dtype
+        )
+
         # LayerNorm
-        self.LayerNorm_weight = weights.LayerNorm.weight.allocate(
-            DType.float32
-        ).cast(dtype)
-        self.LayerNorm_bias = weights.LayerNorm.bias.allocate(
-            DType.float32
-        ).cast(dtype)
-        
-        self.layer_norm_eps = config.layer_norm_eps if hasattr(config, 'layer_norm_eps') else 1e-12
+        self.LayerNorm_weight = weights.LayerNorm.weight.allocate(DType.float32).cast(dtype)
+        self.LayerNorm_bias = weights.LayerNorm.bias.allocate(DType.float32).cast(dtype)
+
+        self.layer_norm_eps = config.layer_norm_eps if hasattr(config, "layer_norm_eps") else 1e-12
 
     def __call__(
         self,
         input_ids: TensorValue,
     ) -> TensorValue:
         """Forward pass.
-        
+
         Args:
             input_ids: Token IDs, shape [batch_size, seq_len]
-            
+
         Returns:
             Embeddings, shape [batch_size, seq_len, hidden_size]
         """
         # Get word embeddings
         embeddings = ops.gather(self.word_embeddings, input_ids, axis=0)
-        
+
         # Create position IDs (0, 1, 2, ..., seq_len-1)
         seq_length = input_ids.shape[1]
         position_ids = ops.range(
@@ -70,7 +64,7 @@ class DistilBertEmbeddings(Module):
             dtype=DType.int64,
             device=input_ids.device,
         )
-        
+
         # Expand position_ids to match batch size
         # Shape: [1, seq_len] -> [batch_size, seq_len]
         position_ids = ops.unsqueeze(position_ids, 0)
@@ -78,13 +72,13 @@ class DistilBertEmbeddings(Module):
             position_ids,
             ("batch_size", seq_length),
         )
-        
+
         # Get position embeddings
         position_embeddings = ops.gather(self.position_embeddings, position_ids, axis=0)
-        
+
         # Add embeddings (no token type embeddings for DistilBERT)
         embeddings = embeddings + position_embeddings
-        
+
         # Layer normalization
         embeddings = ops.layer_norm(
             embeddings,
@@ -92,5 +86,5 @@ class DistilBertEmbeddings(Module):
             self.LayerNorm_bias,
             self.layer_norm_eps,
         )
-        
+
         return embeddings
